requirements.txt
ccxt
pytz
python-dotenv
aiohttp
fastapi
uvicorn
jinja2
PyQt6